requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
aiohttp>=3.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
thefuzz>=0.22.0
python-Levenshtein>=0.25.0
